backend/requirements.txt
# Flask 核心
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0

# 数据库
pymysql==1.1.0
cryptography==41.0.7

# 工具
python-dotenv==1.0.0
marshmallow==3.20.1
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# AI 集成 (可选，用于调用AI API)
openai>=1.10.0
requests==2.31.0
